pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tcp-sdk"
version = "1.0.29"
description = "Python SDK to query The Cross Product API."
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name="Eric Bazan", email="eric.bazan@thecrossproduct.com"},
    {name="Kourosh Beroukhim", email="kourosh.beroukhim@thecrossproduct.com"},
    {name="Theodore Chabardes", email="theodore.chabardes@thecrossproduct.com"},
    {name="David Duque", email="david.duque@thecrossproduct.com"},
    {name="Leonardo Gigli", email="leonardo.gigli@thecrossproduct.com"},
    {name="Benjamin Grappe", email="benjamin.grappe@thecrossproduct.com"},
    {name="Andres Serna", email="andres.serna@thecrossproduct.com"},
    {name="Elise Vi Nhu Ba", email="elise.vinhuba@thecrossproduct.com"}
]
dependencies = [
    "requests-oauthlib>=2.0.0",
    "requests>=2.32.3",
    "slumber>=0.7.1",
    "python-socketio>=5.11.4",
    "websocket-client>=1.8.0",
    "ipython>=8.29.0",
    "click>=8.1.7",
    "pyyaml",
]

[dependency-groups]
dev = [
    "ipython>=8.29.0"
]
test = [
    "google-api-python-client>=2.153.0",
    "google-auth-oauthlib>=1.2.1",
    "pytest>=8.3.3"
]

[tool.hatch.build.targets.wheel]
packages = ["tcp"]
